move template-related code to its own file
