+++ cluster-autoscaler/cloudprovider/azure/azure_scale_set_test.go
@@ -23,7 +23,6 @@ import (
 	"time"
 
 	apiv1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/api/resource"
 	"k8s.io/autoscaler/cluster-autoscaler/cloudprovider"
 	"k8s.io/legacy-cloud-providers/azure/clients/vmssclient/mockvmssclient"
 	"k8s.io/legacy-cloud-providers/azure/clients/vmssvmclient/mockvmssvmclient"
@@ -470,90 +469,4 @@ func TestTemplateNodeInfo(t *testing.T) {
 	assert.NoError(t, err)
 	assert.NotNil(t, nodeInfo)
 	assert.NotEmpty(t, nodeInfo.Pods)
-}
-func TestExtractLabelsFromScaleSet(t *testing.T) {
-	expectedNodeLabelKey := "zip"
-	expectedNodeLabelValue := "zap"
-	extraNodeLabelValue := "buzz"
-	blankString := ""
-
-	tags := map[string]*string{
-		fmt.Sprintf("%s%s", nodeLabelTagName, expectedNodeLabelKey): &expectedNodeLabelValue,
-		"fizz": &extraNodeLabelValue,
-		"bip":  &blankString,
-	}
-
-	labels := extractLabelsFromScaleSet(tags)
-	assert.Len(t, labels, 1)
-	assert.Equal(t, expectedNodeLabelValue, labels[expectedNodeLabelKey])
-}
-
-func TestExtractTaintsFromScaleSet(t *testing.T) {
-	noScheduleTaintValue := "foo:NoSchedule"
-	noExecuteTaintValue := "bar:NoExecute"
-	preferNoScheduleTaintValue := "fizz:PreferNoSchedule"
-	noSplitTaintValue := "some_value"
-	blankTaintValue := ""
-	regularTagValue := "baz"
-
-	tags := map[string]*string{
-		fmt.Sprintf("%s%s", nodeTaintTagName, "dedicated"):                          &noScheduleTaintValue,
-		fmt.Sprintf("%s%s", nodeTaintTagName, "group"):                              &noExecuteTaintValue,
-		fmt.Sprintf("%s%s", nodeTaintTagName, "app"):                                &preferNoScheduleTaintValue,
-		fmt.Sprintf("%s%s", nodeTaintTagName, "k8s.io_testing_underscore_to_slash"): &preferNoScheduleTaintValue,
-		"bar": &regularTagValue,
-		fmt.Sprintf("%s%s", nodeTaintTagName, "blank"):   &blankTaintValue,
-		fmt.Sprintf("%s%s", nodeTaintTagName, "nosplit"): &noSplitTaintValue,
-	}
-
-	expectedTaints := []apiv1.Taint{
-		{
-			Key:    "dedicated",
-			Value:  "foo",
-			Effect: apiv1.TaintEffectNoSchedule,
-		},
-		{
-			Key:    "group",
-			Value:  "bar",
-			Effect: apiv1.TaintEffectNoExecute,
-		},
-		{
-			Key:    "app",
-			Value:  "fizz",
-			Effect: apiv1.TaintEffectPreferNoSchedule,
-		},
-		{
-			Key:    "k8s.io/testing/underscore/to/slash",
-			Value:  "fizz",
-			Effect: apiv1.TaintEffectPreferNoSchedule,
-		},
-	}
-
-	taints := extractTaintsFromScaleSet(tags)
-	assert.Len(t, taints, 4)
-	assert.Equal(t, makeTaintSet(expectedTaints), makeTaintSet(taints))
-}
-
-func TestExtractAllocatableResourcesFromScaleSet(t *testing.T) {
-	tags := map[string]*string{
-		fmt.Sprintf("%s%s", nodeResourcesTagName, "cpu"):               to.StringPtr("100m"),
-		fmt.Sprintf("%s%s", nodeResourcesTagName, "memory"):            to.StringPtr("100M"),
-		fmt.Sprintf("%s%s", nodeResourcesTagName, "ephemeral-storage"): to.StringPtr("20G"),
-	}
-
-	labels := extractAllocatableResourcesFromScaleSet(tags)
-
-	assert.Equal(t, resource.NewMilliQuantity(100, resource.DecimalSI).String(), labels["cpu"].String())
-	expectedMemory := resource.MustParse("100M")
-	assert.Equal(t, (&expectedMemory).String(), labels["memory"].String())
-	expectedEphemeralStorage := resource.MustParse("20G")
-	assert.Equal(t, (&expectedEphemeralStorage).String(), labels["ephemeral-storage"].String())
-}
-
-func makeTaintSet(taints []apiv1.Taint) map[apiv1.Taint]bool {
-	set := make(map[apiv1.Taint]bool)
-	for _, taint := range taints {
-		set[taint] = true
-	}
-	return set
 }
